server/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.4
sqlalchemy==2.0.15
mysql-connector-python==8.0.30
passlib[bcrypt]==1.7.4
python-jose==3.3.0
PyPDF2==3.0.1
fpdf==1.7.2
aiohttp==3.8.4
httpx==0.24.0
gemini-api==1.0.0  # Replace with actual Gemini API package if available
